+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- new `file store test on azure blob storage to employ data contract as storage policy` ([#1227](https://github.com/datacontract/datacontract-cli/issues/1227))
 
 ## [0.12.5] - 2026-05-30